右トリガーの入力が正常に動作しない不具合修正

+++ DXLibProject_Start/Input/Device/GamePad/InputTrigger.cpp
@@ -8,6 +8,6 @@ namespace {
 InputTrigger::InputTrigger(int input)
 {
 	// 入力量を計算
-	m_inputRatio = static_cast<float>(input / kTriggerInputMax);
+	m_inputRatio = static_cast<float>(input) / static_cast<float>(kTriggerInputMax);
 }
